upgrade deps; new lint rules
