+++ package.json
@@ -23,16 +23,16 @@
   },
   "dependencies": {
     "@codesandbox/sandpack-react": "2.19.9",
-    "@leanstacks/react-common": "1.0.0",
+    "@leanstacks/react-common": "1.1.0",
     "@react-spring/web": "9.7.5",
-    "@tanstack/react-query": "5.59.15",
-    "@tanstack/react-query-devtools": "5.59.15",
+    "@tanstack/react-query": "5.59.16",
+    "@tanstack/react-query-devtools": "5.59.16",
     "@tanstack/react-table": "8.20.5",
     "axios": "1.7.7",
     "classnames": "2.5.1",
     "dayjs": "1.11.13",
     "formik": "2.4.6",
-    "i18next": "23.16.2",
+    "i18next": "23.16.4",
     "i18next-browser-languagedetector": "8.0.0",
     "lodash": "4.17.21",
     "qs": "6.13.0",
@@ -59,9 +59,9 @@
     "autoprefixer": "10.4.20",
     "eslint": "8.57.1",
     "eslint-plugin-react-hooks": "5.0.0",
-    "eslint-plugin-react-refresh": "0.4.13",
+    "eslint-plugin-react-refresh": "0.4.14",
     "jsdom": "25.0.1",
-    "msw": "2.5.0",
+    "msw": "2.5.2",
     "postcss": "8.4.47",
     "prettier": "3.3.3",
     "prettier-plugin-tailwindcss": "0.6.8",
