+++ src/providers/AxiosProvider.tsx
@@ -1,7 +1,8 @@
-import React, { PropsWithChildren, useEffect, useState } from 'react';
-import axios, { AxiosError, AxiosInstance, InternalAxiosRequestConfig } from 'axios';
+import { PropsWithChildren, useEffect, useState } from 'react';
+import axios, { AxiosError, InternalAxiosRequestConfig } from 'axios';
 
-import { AuthContextValue } from './AuthProvider';
+import { AxiosContext, customAxios } from './AxiosContext';
+import { AuthContextValue } from './AuthContext';
 import { useAuth } from 'hooks/useAuth';
 
 /**
@@ -27,8 +28,8 @@ const authRequestInterceptor = async (
 /**
  * An Axios response interceptor called for responses in error. If the http status
  * code is `401`, attempts to refresh the authentication tokens and retry the request.
- * @param error {AxiosError} error - The AxiosError instance.
- * @param authContext {AuthContextValue} authContext - The `AuthContextValue` containing
+ * @param {AxiosError} error - The AxiosError instance.
+ * @param {AuthContextValue} authContext - The `AuthContextValue` containing
  * the current user authentication state.
  * @returns {AxiosResponse} An AxiosResponse representing the retried request.
  */
@@ -56,27 +57,12 @@ const notAuthenticatedErrorInterceptor = async (
 };
 
 /**
- * Custom `Axios` instance.
- */
-const customAxios = axios.create({
-  headers: {
-    Accept: 'application/json',
-    'Content-Type': 'application/json',
-  },
-});
-
-/**
- * The `AxiosContext` instance.
- */
-export const AxiosContext = React.createContext<AxiosInstance>(customAxios);
-
-/**
  * The `AxiosContextProvider` React component creates, maintains, and provides
  * access to the `AxiosContext` value.
  * @param {PropsWithChildren} props - Component properties, `PropsWithChildren`.
  * @returns {JSX.Element} JSX
  */
-const AxiosContextProvider = ({ children }: PropsWithChildren) => {
+const AxiosContextProvider = ({ children }: PropsWithChildren): JSX.Element => {
   const [isReady, setIsReady] = useState(false);
   const authContext = useAuth();
